+++ shortcodes/Footer/FooterTwoShortcode.php
@@ -1,0 +1,43 @@
+<?php
+/**
+ * This file is part of the Gravstrap plugin and it is distributed
+ * under the MIT License. To use this application you must leave
+ * intact this copyright notice.
+ *
+ * Copyright (c) Giansimon Diblas <info@diblas.net>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * For extra documentation and help please visit http://diblas.net
+ *
+ * @license    MIT License
+ *
+ */
+
+namespace Grav\Plugin\Shortcodes;
+
+
+/**
+ * Class FooterTwo handles a basic footer module with license
+ *
+ * @author Giansimon Diblas
+ */
+class FooterTwoShortcode extends FooterOneShortcode
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function shortcodeName()
+    {
+        return 'gravstrap-footer-two';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function template()
+    {
+        return 'footer/footer_two.html.twig';
+    }
+}